+++ src/themes/bootstrap4.js
@@ -380,6 +380,29 @@ class ThemeBootstrap4 extends Theme {
     control.container.classList.remove('form-group')
   }
 
+  getSwitcherSelect (config) {
+    const control = super.getSwitcherSelect(config)
+    control.input.classList.add('form-control-sm')
+    return control
+  }
+
+  getSwitcherModal (config) {
+    const control = super.getSwitcherModal(config)
+    control.container.classList.add('ml-1')
+    control.trigger.classList.add('badge', 'badge-secondary')
+    control.dialogBody.classList.add('btn-group', 'btn-group-vertical', 'w-100')
+    control.optionButtons.forEach(btn => {
+      btn.classList.add('btn', 'btn-secondary')
+    })
+    return control
+  }
+
+  setSwitcherOptionActive (btn, active) {
+    super.setSwitcherOptionActive(btn, active)
+    btn.classList.toggle('btn-primary', active)
+    btn.classList.toggle('btn-secondary', !active)
+  }
+
   adaptForTableMultipleControl (control, td) {
     super.adaptForTableMultipleControl(control, td)
     control.container.classList.remove('mb-3')